refactor(callbacks): extract Backtest tab helpers to _callbacks_backtest.py
Step 9 of the callbacks.py decomposition (issue #87). Final per-tab
extraction — every demand / forecast / backtest / overview helper now
lives in a per-tab module. callbacks.py is reduced to the Redis-fetch
fast-path entry point plus ``register_callbacks``.

## What moves

Seven helpers covering the Backtest tab surface:

* ``_normalize_backtest_exog_mode`` — coerce optional UI input into
  one of {``oracle_exog``, ``forecast_exog``}.
* ``_describe_exog_mode`` — human-readable label for chart titles.
* ``_build_forecast_exog_fold`` — production-like exogenous weather
  for a backtest fold, with Redis snapshot + climatology fallback.
* ``_backtest_tab_from_redis`` — Redis fast path that builds the
  entire Backtest tab from the scoring job's hourly
  ``wattcast:backtest:{exog_mode}:{region}:{horizon_hours}`` payload.
* ``_predict_single_fold`` — train+predict for one fold of one model.
* ``_ensemble_fold`` — equal-weight ensemble across all models for
  one fold. Uses uniform averaging (not 1/MAPE) to avoid data leakage
  — see the docstring for the leakage argument.
* ``_run_backtest_for_horizon`` — walk-forward expanding-window
  cross-validation orchestrator with 3-tier cache (in-memory →
  SQLite → train fresh) and warming gate for production.

## Why these stay together (vs splitting fast-path from compute path)

``_backtest_tab_from_redis`` (the tab's read path) and
``_run_backtest_for_horizon`` (the v1 compute path that the training
Cloud Run Job invokes) are the two entry points the tab uses. The
remaining five helpers are private to one or the other. Splitting
fast-path from compute-path would leave the leakage-control rationale
(``_ensemble_fold``'s uniform-averaging docstring) and the exog-mode
normalization wiring straddling two modules. Keeping them together
preserves cohesion.

## Public-import surface preserved

callbacks.py re-imports each helper by name. Tests using ``from
components.callbacks import _run_backtest_for_horizon`` etc continue
to resolve via the re-export shim. ``# noqa: F401`` markers on all
imports since the helpers are no longer called inside callbacks.py
itself (``register_callbacks`` wires them up, but the actual call
sites for the backtest functions are either in tests or in
``register_callbacks`` Dash decorators that wrap the helpers).

Re-added ``_BACKTEST_CACHE`` and ``_CACHE_VERSION`` to the shared
re-export block with ``# noqa`` markers — the test fixture
``_clear_module_caches`` calls ``cb._BACKTEST_CACHE.clear()`` and
v1-paths tests reference ``cb._CACHE_VERSION``.

## Test patches updated

Bulk-rewired ``@patch`` decorators across three test files:

* ``test_redis_fast_paths.py::TestBacktestTabFromRedis`` — 8 ``redis_get``
  patches rewired to ``components._callbacks_backtest.redis_get``.
* ``test_callbacks_helpers.py`` (TestPredictSingleFold +
  TestEnsembleFold + TestRunBacktestForHorizon, lines 1104-1743) —
  patches of ``_predict_single_fold``, ``_ensemble_fold``,
  ``redis_get``, ``_BACKTEST_CACHE``, ``REQUIRE_REDIS``, and
  ``_compute_data_hash`` rewired to the new module path.
* ``test_callbacks_v1_paths.py`` (same three classes, lines 441-645)
  — same pattern.

Integration test ``TestBacktestWarming`` patches
``REQUIRE_REDIS`` on both ``components.callbacks`` AND
``components._callbacks_backtest`` (the dual-patch pattern established
in PR #107 for the Forecast extraction).

## Size impact

callbacks.py: 2,924 → 2,255 (–669 lines, –22.9%)
_callbacks_backtest.py: new, 775 lines

**callbacks.py crosses below 2,300 lines.** Remaining ~2,255 is mostly
``register_callbacks`` itself (~2,000 lines) plus ``_load_data_from_redis``
and the US Grid re-export tail.

## Cumulative progress

* PR #98 (shared):       7,167 → 7,051 (–116)
* PR #99 (us_grid):      7,051 → 6,389 (–662)
* PR #100 (models):      6,389 → 6,183 (–206)
* PR #101 (alerts):      6,183 → 5,978 (–205)
* PR #102 (generation):  5,978 → 5,802 (–176)
* PR #103 (weather):     5,802 → 5,668 (–134)
* PR #104 (overview 7a): 5,668 → 5,369 (–299)
* PR #105 (overview 7b): 5,369 → 4,686 (–683)
* PR #106 (overview 7c): 4,686 → 3,750 (–936)
* PR #107 (forecast):    3,750 → 2,924 (–826)
* this PR (backtest):    2,924 → 2,255 (–669)

Total: −4,912 lines from the monolith across 11 PRs (−68.5%).

## What's left

Just one structural move remaining: split ``register_callbacks`` into
per-tab ``register_X_callbacks(app)`` functions. That's the final
step that lets the per-tab modules own their own callback registration
end-to-end.

## Tests

All 1,573 unit + integration tests pass locally. Ruff clean.
